Restructure and refactoring of project
Removed backend controller function which could be merged with another function. Changed interfaces to 	ype when I didn't require the extra functionality. Imported using * as someName for my backend routes as it was getting quite long. Changed a post request to a patch request. Renamed some interfaces to remove any I or Interface mentioned in the naming. Added layout routing for the Navbar